moved pngs to public
