+++ apps/main-landing/next.config.ts
@@ -5,9 +5,6 @@ const nextConfig: NextConfig = {
   experimental: {
     optimizePackageImports: ['@idriss-xyz/ui'],
   },
-  images: {
-    disableStaticImages: false,
-  }
 };
 
 // eslint-disable-next-line import/no-default-export
