+++ apps/main-landing/public/hero-background-lines.png
