--- apps/main-landing/src/components/hero-section/components/background-lines/background-lines.tsx
@@ -1,14 +1,0 @@
-import Image from 'next/image';
-
-import lines from './lines.png';
-
-export const BackgroundLines = () => {
-  return (
-    <Image
-      priority
-      src={lines}
-      className="absolute inset-x-20 top-0 hidden opacity-40 lg:block"
-      alt=""
-    />
-  );
-};
