+++ apps/main-landing/src/components/hero-section/components/index.ts
@@ -1,3 +1,3 @@
-export { BackgroundLines } from './background-lines/background-lines';
-export { TokensFlowImage } from './tokens-flow/tokens-flow-image';
+export { BackgroundLines } from './background-lines';
+export { TokensFlowImage } from './tokens-flow-image';
 export { CollaborationsShowcase } from './collaborations-showcase';
