+++ apps/main-landing/src/components/hero-section/components/tokens-flow-image.tsx
@@ -1,12 +1,12 @@
 import Image from 'next/image';
 
-import tokensFlow from './tokens-flow.png';
-
 export const TokensFlowImage = () => {
   return (
     <Image
       priority
-      src={tokensFlow}
+      width={1985}
+      height={1240}
+      src={'/hero-tokens-flow.png'}
       className="z-0 -mt-[22%] w-full min-w-[500px] md:-mt-[18%] lg:-mt-[15%]"
       alt=""
     />
