+++ apps/main-landing/tsconfig.json
@@ -18,7 +18,6 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-     "**/*.png"
   ],
   "exclude": [
     "node_modules"
